Project_02/requirements.txt
flask
flask-sqlalchemy
flask-user==1.0.2.2
scikit-learn
numpy==1.26.1
beautifulsoup4==4.12.2
requests==2.31.0
urllib3==2.0.7